backend/auth/requirements.txt
psycopg2-binary>=2.9.9
bcrypt>=4.1.2
